# Updated Heap algorithm