--- a/divide_and_conquer/heaps_algorithm_iterative.py
+++ b/divide_and_conquer/heaps_algorithm_iterative.py
@@ -6,6 +6,7 @@ More information:
 https://en.wikipedia.org/wiki/Heap%27s_algorithm.
 """
 
+
 def heaps(arr: list) -> list:
     """
     Pure python implementation of the Heap's algorithm (recursive version),