@cclauss 2023-10-01
What are the advantages of the proposed changes?  Are they faster?  Can we prove that with a `timeit` benchmark.
* https://github.com/search?q=repo%253ATheAlgorithms%252FPython%20timeit
* https://docs.python.org/3/library/timeit.html