@abhinavjha98 2023-10-02
# Proposed Changes to Heap's Algorithm

## Advantages

### Simplification
The revised version introduces a more self-contained approach by passing the result list (`res`) directly to the recursive function. This decouples the function from outer-scope variables, enhancing its modularity and readability.

### Efficiency
Relocating the `len(arr) <= 1` check outside the recursive function ensures it's executed only once. This offers a performance boost, especially for longer input lists, by reducing unnecessary checks.

## `timeit` Benchmark Results

| Test List       | Original Algorithm (s) | Proposed Algorithm (s) |
|-----------------|------------------------|------------------------|
| `[1, 2, 3, 4]`  | 0.111138824            | 0.067861666            |

For the input list `[1, 2, 3, 4]`, the proposed algorithm is notably faster than the original, showcasing its better scalability with larger inputs.

## Conclusion

The proposed changes, while introducing enhanced code clarity, also offer improved performance for more extensive input lists, making a compelling case for this revision.